Prepare test directory.
The file __init__.py is empty and serves to declare the test suite as a package. While this is not strictly necessary, it allows your test suite to mirror the source layout of the package under test, even when modules in different parts of the source tree have the same name. Furthermore, it gives you the option to import modules from within your tests package.

The file test_console.py contains a test case for the console module, which checks whether the program exits with a status code of zero.